+++ src/main/java/br/pucrio/les/esg_token_backend/resources/users/repository/IUserRepository.java
@@ -1,6 +1,8 @@
 package br.pucrio.les.esg_token_backend.resources.users.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -8,6 +10,6 @@ import br.pucrio.les.esg_token_backend.resources.users.model.User;
 
 @Repository
 @Transactional
-public interface IUserRepository extends JpaRepository<User, Long> {
-
+public interface IUserRepository extends CrudRepository<User, Long> {
+    Optional<User> findByEmail(String email);
 }