+++ src/main/java/br/pucrio/les/esg_token_backend/resources/users/services/UserService.java
@@ -1,0 +1,36 @@
+package br.pucrio.les.esg_token_backend.resources.users.services;
+
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import br.pucrio.les.esg_token_backend.resources.users.model.User;
+import br.pucrio.les.esg_token_backend.resources.users.repository.IUserRepository;
+
+@Service
+public class UserService implements IUserService {
+
+    @Autowired
+    private IUserRepository userRepository;
+
+    public UserService(IUserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public Optional<User> findById(Long id) {
+        return this.userRepository.findById(id);
+    }
+
+    @Override
+    public Optional<User> findByEmail(String email) {
+        return this.userRepository.findByEmail(email);
+    }
+
+    @Override
+    public User create(User user) {
+        return this.userRepository.save(user);
+    }
+
+}